+++ src/util/vgmstream_limits.h
